requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
Bottleneck @ file:///C:/miniconda3/conda-bld/bottleneck_1761938191468/work
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592920106/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593150425/work
narwhals==2.13.0
numexpr @ file:///C:/miniconda3/conda-bld/numexpr_1762165733453/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1763980698946/work/dist/numpy-2.3.5-cp311-cp311-win_amd64.whl#sha256=e3ba89fff46662b034bf265cf2c543d64ef34c25de2a336da7724a746cb8dc4e
openai==2.8.1
packaging==25.0
pandas @ file:///C:/miniconda3/conda-bld/pandas_1762332399011/work/dist/pandas-2.3.3-cp311-cp311-win_amd64.whl#sha256=6a3251bc4b4b7b67e557583d27c46e59dd38a774af7276284aaf0d4b219ca605
pillow==12.0.0
protobuf==6.33.1
pyarrow @ file:///C:/miniconda3/conda-bld/pyarrow_1759833600682/work/python
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.2.1
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.5.0
watchdog==6.0.0
